+++ src/main/java/me/despical/tntrun/handlers/PlaceholderManager.java
@@ -45,7 +45,7 @@ public class PlaceholderManager extends PlaceholderExpansion {
 	}
 
 	public String getVersion() {
-		return "1.0.7";
+		return "1.0.8";
 	}
 
 	public String onPlaceholderRequest(Player player, String id) {