+++ src/main/java/me/despical/tntrun/handlers/sign/SignManager.java
@@ -36,6 +36,7 @@ import org.bukkit.block.Block;
 import org.bukkit.block.Sign;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.Action;
 import org.bukkit.event.block.BlockBreakEvent;
@@ -62,7 +63,6 @@ public class SignManager implements Listener {
 
 	public SignManager(Main plugin) {
 		this.plugin = plugin;
-		FileConfiguration config = ConfigUtils.getConfig(plugin, "messages");
 
 		gameStateToString.put(ArenaState.WAITING_FOR_PLAYERS, plugin.getChatManager().colorMessage("Signs.Game-States.Waiting"));
 		gameStateToString.put(ArenaState.STARTING, plugin.getChatManager().colorMessage("Signs.Game-States.Starting"));
@@ -71,12 +71,9 @@ public class SignManager implements Listener {
 		gameStateToString.put(ArenaState.RESTARTING, plugin.getChatManager().colorMessage("Signs.Game-States.Restarting"));
 		gameStateToString.put(ArenaState.INACTIVE, plugin.getChatManager().colorMessage("Signs.Game-States.Inactive"));
 
-		signLines = config.getStringList("Signs.Lines");
+		signLines = plugin.getChatManager().getStringList("Signs.Lines");
 
 		plugin.getServer().getPluginManager().registerEvents(this, plugin);
-
-		loadSigns();
-		updateSignScheduler();
 	}
 
 	@EventHandler
@@ -170,7 +167,7 @@ public class SignManager implements Listener {
 		e.getPlayer().sendMessage(plugin.getChatManager().getPrefix() + ChatColor.RED + "Couldn't remove sign from configuration! Please do this manually!");
 	}
 
-	@EventHandler
+	@EventHandler(ignoreCancelled = true, priority = EventPriority.HIGH)
 	public void onJoinAttempt(PlayerInteractEvent e) {
 		ArenaSign arenaSign = getArenaSignByBlock(e.getClickedBlock());
 
@@ -228,7 +225,6 @@ public class SignManager implements Listener {
 				if (loc.getBlock().getState() instanceof Sign) {
 					arenaSigns.add(new ArenaSign((Sign) loc.getBlock().getState(), ArenaRegistry.getArena(path)));
 				} else {
-					arenaSigns.remove(getArenaSignByBlock(loc.getBlock()));
 					Debugger.debug(Level.WARNING, "Block at location {0} for arena {1} not a sign", loc, path);
 				}
 			}
@@ -237,7 +233,7 @@ public class SignManager implements Listener {
 		Debugger.debug("Sign load event finished took {0} ms", System.currentTimeMillis() - start);
 	}
 
-	private void updateSignScheduler() {
+	public void updateSigns() {
 		Bukkit.getScheduler().runTaskTimer(plugin, () -> {
 			Debugger.performance("SignUpdate", "[PerformanceMonitor] [SignUpdate] Updating signs");